+++ tests/testthat/test-cbbinom.R
@@ -6,6 +6,8 @@ test_size <- 10
 test_alpha <- 2
 test_beta <- 4
 test_delta <- 1e-6
+test_tol <- 1e-6
+test_max_iter <- 10000L
 
 # Inputs
 test_x <- 5
@@ -43,6 +45,21 @@ testthat::test_that(
       qcbbinom(p = test_val, size = test_size, alpha = test_alpha, beta = test_beta),
       test_x
     )
+    # Verify that p, root_tol and root_max_iter are not changed
+    test_val_log <- test_val_log_orig <- log(test_val)
+    test_tol_orig <- test_tol
+    test_max_iter_orig <- test_max_iter
+    testthat::expect_equal(
+      cpp_qcbbinom(p = test_val_log, size = test_size,
+                   alpha = test_alpha, beta = test_beta,
+                   lower_tail = TRUE, log_p = TRUE,
+                   p_tol = test_tol, p_max_iter = test_max_iter,
+                   root_tol = test_tol, root_max_iter = test_max_iter),
+      test_x
+    )
+    testthat::expect_identical(test_val_log, test_val_log_orig)
+    testthat::expect_identical(test_tol, test_tol_orig)
+    testthat::expect_identical(test_max_iter, test_max_iter_orig)
   }
 )
 
@@ -68,8 +85,8 @@ testthat::test_that(
       gen_hypergeo(U = U,
                    L = L,
                    x = 1,
-                   tol = 1e-6,
-                   max_iter = 10000L,
+                   tol = test_tol,
+                   max_iter = test_max_iter,
                    check_mode = TRUE,
                    log = FALSE),
       101/5460