+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Fix a bug in `gen_hypergeo()` that may change input `U` and `L`.
 
+* Fix a bug in `cpp_qcbbinom()` that may change input `p`, `tol` and `max_iter`.
+
 
 # cbbinom 0.1.0